Disable automatic speech synthesis
The Bookworm release added support for installing using a software
speech synthesis.  Unfortunately, the "ontimeout" method they used to
add this to the BIOS installer means that it was activated for this
automated installer - and it required user input to proceed to the
network install.

The speech synthesis is still present in this installer, but it now
needs a sighted person to navigate to the advanced menu to turn it on.